move part of data loading logic from project-base into packages
- MultipleSearchQuery + Factory moved to the FE API package
- ProductBatchLoader:load*ByIds moved to the FE API package (the new class is now a parent of ProductBatchLoader from project-base)
- ProductBatchLoader::getTotalByBatchLoadDataId moved to the FE API package (the new class is now a parent of ProductBatchLoader from project-base)
- ProductElasticsearchProvider::getBatched*ByProductIds moved to a new class "ProductElasticsearchBatchProvider" in FE API package
- ProductElasticsearchRepository::getBatchedProductsAndTotalsByFilterQueries moved to a new class "ProductElasticsearchBatchRepository" in FE API package
- ProductElasticsearchRepository::PRODUCTS_KEY and TOTALS_KEY constants moved to a new class "ProductElasticsearchBatchRepository" in FE API package
