+++ packages/frontend-api/composer.json
@@ -22,9 +22,11 @@
     },
     "require": {
         "php": "^8.1",
+        "elasticsearch/elasticsearch": "^7.6.1",
         "lcobucci/jwt": "^4.1.5",
         "overblog/graphql-bundle": "^1.0",
         "overblog/graphiql-bundle": "^0.3",
+        "overblog/dataloader-bundle": "^0.6.0",
         "shopsys/form-types-bundle": "14.0.x-dev",
         "shopsys/framework": "14.0.x-dev",
         "shopsys/migrations": "14.0.x-dev",
