+++ packages/frontend-api/src/Resources/config/services.yaml
@@ -5,7 +5,7 @@ services:
         public: false
 
     Shopsys\FrontendApiBundle\:
-        resource: '../../**/*{Facade,Factory,Mapper,Mutation,Query,Repository,Subscriber,Validator}.php'
+        resource: '../../**/*{Facade,Factory,Mapper,Mutation,Query,Repository,Subscriber,Validator,Provider}.php'
 
     resolverMaps:
         namespace: Shopsys\FrontendApiBundle\
