+++ project-base/app/src/Model/Product/Search/ProductElasticsearchRepository.php
@@ -4,12 +4,8 @@ declare(strict_types=1);
 
 namespace App\Model\Product\Search;
 
-use App\Component\Elasticsearch\MultipleSearchQueryFactory;
 use App\Model\Product\Filter\ProductFilterData;
 use Doctrine\ORM\QueryBuilder;
-use Elasticsearch\Client;
-use Shopsys\FrameworkBundle\Component\Elasticsearch\IndexDefinitionLoader;
-use Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ProductIndex;
 use Shopsys\FrameworkBundle\Model\Product\Search\ProductElasticsearchRepository as BaseProductElasticsearchRepository;
 
 /**
@@ -19,34 +15,14 @@ use Shopsys\FrameworkBundle\Model\Product\Search\ProductElasticsearchRepository 
  * @method \Shopsys\FrameworkBundle\Model\Product\Search\ProductsResult getSortedProductsResultByFilterQuery(\App\Model\Product\Search\FilterQuery $filterQuery)
  * @method int getProductsCountByFilterQuery(\App\Model\Product\Search\FilterQuery $filterQuery)
  * @method array getProductsByFilterQuery(\App\Model\Product\Search\FilterQuery $filterQuery)
+ * @method __construct(\Elasticsearch\Client $client, \App\Model\Product\Search\ProductElasticsearchConverter $productElasticsearchConverter, \App\Model\Product\Search\FilterQueryFactory $filterQueryFactory, \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexDefinitionLoader $indexDefinitionLoader)
  */
 class ProductElasticsearchRepository extends BaseProductElasticsearchRepository
 {
-    public const PRODUCTS_KEY = 'products';
-
-    public const TOTALS_KEY = 'totals';
-
-    /**
-     * @param \Elasticsearch\Client $client
-     * @param \App\Model\Product\Search\ProductElasticsearchConverter $productElasticsearchConverter
-     * @param \App\Model\Product\Search\FilterQueryFactory $filterQueryFactory
-     * @param \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexDefinitionLoader $indexDefinitionLoader
-     * @param \App\Component\Elasticsearch\MultipleSearchQueryFactory $multipleSearchQueryFactory
-     */
-    public function __construct(
-        Client $client,
-        ProductElasticsearchConverter $productElasticsearchConverter,
-        FilterQueryFactory $filterQueryFactory,
-        IndexDefinitionLoader $indexDefinitionLoader,
-        private MultipleSearchQueryFactory $multipleSearchQueryFactory,
-    ) {
-        parent::__construct($client, $productElasticsearchConverter, $filterQueryFactory, $indexDefinitionLoader);
-    }
-
     /**
      * {@inheritdoc}
      */
-    protected function extractTotalCount(array $result): int
+    public function extractTotalCount(array $result): int
     {
         return (int)$result['hits']['total']['value'];
     }
@@ -94,29 +70,5 @@ class ProductElasticsearchRepository extends BaseProductElasticsearchRepository
         }
 
         return $result;
-    }
-
-    /**
-     * @param \App\Model\Product\Search\FilterQuery[] $filterQueries
-     * @return array
-     */
-    public function getBatchedProductsAndTotalsByFilterQueries(array $filterQueries): array
-    {
-        $mSearchQuery = $this->multipleSearchQueryFactory->create(ProductIndex::getName(), $filterQueries);
-        $result = $this->client->msearch($mSearchQuery->getQuery());
-
-        $keys = array_keys($filterQueries);
-        $products = [];
-        $totals = [];
-
-        foreach ($result['responses'] as $index => $response) {
-            $products[$keys[$index]] = $this->extractHits($response);
-            $totals[$keys[$index]] = $this->extractTotalCount($response);
-        }
-
-        return [
-            self::PRODUCTS_KEY => $products,
-            self::TOTALS_KEY => $totals,
-        ];
     }
 }
